Responding to feedback from PR
